move getSegmentEl() from InfoBubble to generic module
